<fix>:使用std::array来替换简单的指针

+++ src/gui/lon_widget/listwidget.hpp
@@ -139,7 +139,7 @@ public:
         return widget_list_[index];
     }
 
-    int count() {
+    size_t count() {
         return widget_list_.size();
     }
